fault_controller -> picknik_reset_fault_controller
This commit does two things:

1) renames fault_controller to reset_fault_controller
2) prefixes with picknik_

The first change, is to be more specific what this controller is used
for.

The second change is because we want to move this controller into
ros2_controllers and when that is complete we can drop the picknik_ and
depricate this version allowing for a transition period.

Signed-off-by: Alex Moriarty <alex.moriarty@picknik.ai>